Comparison of local files working!

+++ pkg/python/python.go
@@ -1,7 +1,7 @@
 package python
 
 import (
-	"fmt"
+	//"fmt"
 	"log"
 	"os"
 	"os/exec"
@@ -26,36 +26,64 @@ func Diff(filename string, options types.Options, config types.Config) string {
 	var head []byte
 	var headTree []byte
 	var err error
-	pythonCmd := config.Commands["python"].Executable
-	switches := append(config.Commands["python"].Switches, filename)
+	var pythonCmd string = config.Commands["python"].Executable
+	var switches []string = config.Commands["python"].Switches
 
-	// Get the AST for the current version of the file
-	cmdCurrentTree := exec.Command(pythonCmd, switches...)
-	currentTree, err = cmdCurrentTree.Output()
-	if err != nil {
-		log.Fatal(err)
-	}
+	if filename == "" {
+		filename = options.Source + " -> " + options.Destination
+		utils.Info.Println("Comparing local files: " + filename)
+		//-- Empty filename compares options.Source and options.Destination --
+		// The prefixes "head" and "current" are slight misnomers here
+		cmdHeadTree := exec.Command(pythonCmd,
+			append(switches, options.Source)...,
+		)
+		headTree, err = cmdHeadTree.Output()
+		if err != nil {
+			utils.Fail.Println("Could not create parse tree for " + options.Source)
+			log.Fatal(err)
+		}
 
-	// Retrieve the HEAD version of the file to a temporary filename
-	cmdHead := exec.Command("git", "show", fmt.Sprintf("HEAD:%s", filename))
-	head, err = cmdHead.Output()
-	if err != nil {
-		log.Fatal(err)
-	}
+		cmdCurrentTree := exec.Command(pythonCmd,
+			append(switches, options.Destination)...,
+		)
+		currentTree, err = cmdCurrentTree.Output()
+		if err != nil {
+			utils.Fail.Println("Could not create parse tree for " + options.Destination)
+			log.Fatal(err)
+		}
+	} else {
+		//-- Comparing a branch/revision to current local file --
+		// Get the AST for the current version of the file
+		cmdCurrentTree := exec.Command(pythonCmd,
+			append(switches, filename)...,
+		)
+		currentTree, err = cmdCurrentTree.Output()
+		if err != nil {
+			log.Fatal(err)
+		}
 
-	tmpfile, err := os.CreateTemp("", "*.py")
-	if err != nil {
-		log.Fatal(err)
-	}
-	tmpfile.Write(head)
-	defer os.Remove(tmpfile.Name()) // clean up
+		// Retrieve the HEAD version of the file to a temporary filename
+		cmdHead := exec.Command("git", "show", options.Source+filename)
+		head, err = cmdHead.Output()
+		if err != nil {
+			log.Fatal(err)
+		}
 
-	// Get the AST for the HEAD version of the file
-	switches = append(config.Commands["python"].Switches, tmpfile.Name())
-	cmdHeadTree := exec.Command(pythonCmd, switches...)
-	headTree, err = cmdHeadTree.Output()
-	if err != nil {
-		log.Fatal(err)
+		tmpfile, err := os.CreateTemp("", "*.py")
+		if err != nil {
+			log.Fatal(err)
+		}
+		tmpfile.Write(head)
+		defer os.Remove(tmpfile.Name()) // clean up
+
+		// Get the AST for the HEAD version of the file
+		cmdHeadTree := exec.Command(pythonCmd,
+			append(switches, tmpfile.Name())...,
+		)
+		headTree, err = cmdHeadTree.Output()
+		if err != nil {
+			log.Fatal(err)
+		}
 	}
 
 	// Make the trees into slightly simpler string representation
